Refactored Cell class into its own script
